+++ Src/FluentAssertions/Formatting/Formatter.cs
@@ -31,6 +31,9 @@ namespace FluentAssertions.Formatting
             new NullValueFormatter(),
             new GuidValueFormatter(),
             new DateTimeOffsetValueFormatter(),
+#if NET6_0_OR_GREATER
+            new DateOnlyValueFormatter(),
+#endif
             new TimeSpanValueFormatter(),
             new Int32ValueFormatter(),
             new Int64ValueFormatter(),