+++ Src/FluentAssertions/Primitives/DateOnlyAssertions.cs
@@ -1,0 +1,519 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using FluentAssertions.Execution;
+
+#if NET6_0_OR_GREATER
+
+namespace FluentAssertions.Primitives
+{
+    /// <summary>
+    /// Contains a number of methods to assert that a <see cref="DateOnly"/> is in the expected state.
+    /// </summary>
+    [DebuggerNonUserCode]
+    public class DateOnlyAssertions : DateOnlyAssertions<DateOnlyAssertions>
+    {
+        public DateOnlyAssertions(DateOnly? value)
+            : base(value)
+        {
+        }
+    }
+
+#pragma warning disable CS0659 // Ignore not overriding Object.GetHashCode()
+#pragma warning disable CA1065 // Ignore throwing NotSupportedException from Equals
+    /// <summary>
+    /// Contains a number of methods to assert that a <see cref="DateOnly"/> is in the expected state.
+    /// </summary>
+    [DebuggerNonUserCode]
+    public class DateOnlyAssertions<TAssertions>
+        where TAssertions : DateOnlyAssertions<TAssertions>
+    {
+        public DateOnlyAssertions(DateOnly? value)
+        {
+            Subject = value;
+        }
+
+        /// <summary>
+        /// Gets the object which value is being asserted.
+        /// </summary>
+        public DateOnly? Subject { get; }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/> is exactly equal to the <paramref name="expected"/> value.
+        /// </summary>
+        /// <param name="expected">The expected value</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> Be(DateOnly expected, string because = "", params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .ForCondition(Subject == expected)
+                .BecauseOf(because, becauseArgs)
+                .FailWith("Expected {context:date} to be {0}{reason}, but found {1}.",
+                    expected, Subject);
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/> is exactly equal to the <paramref name="expected"/> value.
+        /// </summary>
+        /// <param name="expected">The expected value</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> Be(DateOnly? expected, string because = "", params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .ForCondition(Subject == expected)
+                .BecauseOf(because, becauseArgs)
+                .FailWith("Expected {context:date} to be {0}{reason}, but found {1}.",
+                    expected, Subject);
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/> or <see cref="DateOnly"/> is not equal to the <paramref name="unexpected"/> value.
+        /// </summary>
+        /// <param name="unexpected">The unexpected value</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> NotBe(DateOnly unexpected, string because = "",
+            params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .ForCondition(Subject != unexpected)
+                .BecauseOf(because, becauseArgs)
+                .FailWith("Expected {context:date} not to be {0}{reason}, but it is.", unexpected);
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/> or <see cref="DateOnly"/> is not equal to the <paramref name="unexpected"/> value.
+        /// </summary>
+        /// <param name="unexpected">The unexpected value</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> NotBe(DateOnly? unexpected, string because = "",
+            params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .ForCondition(Subject != unexpected)
+                .BecauseOf(because, becauseArgs)
+                .FailWith("Expected {context:date} not to be {0}{reason}, but it is.", unexpected);
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/>  is before the specified value.
+        /// </summary>
+        /// <param name="expected">The <see cref="DateOnly"/>  that the current value is expected to be before.</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> BeBefore(DateOnly expected, string because = "",
+            params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .ForCondition(Subject.HasValue && Subject.Value.CompareTo(expected) < 0)
+                .BecauseOf(because, becauseArgs)
+                .FailWith("Expected {context:date} to be before {0}{reason}, but found {1}.", expected,
+                    Subject);
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/>  is not before the specified value.
+        /// </summary>
+        /// <param name="unexpected">The <see cref="DateOnly"/>  that the current value is not expected to be before.</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> NotBeBefore(DateOnly unexpected, string because = "",
+            params object[] becauseArgs)
+        {
+            return BeOnOrAfter(unexpected, because, becauseArgs);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/>  is either on, or before the specified value.
+        /// </summary>
+        /// <param name="expected">The <see cref="DateOnly"/>  that the current value is expected to be on or before.</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> BeOnOrBefore(DateOnly expected, string because = "",
+            params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .ForCondition(Subject.HasValue && Subject.Value.CompareTo(expected) <= 0)
+                .BecauseOf(because, becauseArgs)
+                .FailWith("Expected {context:date} to be on or before {0}{reason}, but found {1}.", expected,
+                    Subject);
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/>  is neither on, nor before the specified value.
+        /// </summary>
+        /// <param name="unexpected">The <see cref="DateOnly"/>  that the current value is not expected to be on nor before.</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> NotBeOnOrBefore(DateOnly unexpected, string because = "",
+            params object[] becauseArgs)
+        {
+            return BeAfter(unexpected, because, becauseArgs);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/>  is after the specified value.
+        /// </summary>
+        /// <param name="expected">The <see cref="DateOnly"/>  that the current value is expected to be after.</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> BeAfter(DateOnly expected, string because = "",
+            params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .ForCondition(Subject.HasValue && Subject.Value.CompareTo(expected) > 0)
+                .BecauseOf(because, becauseArgs)
+                .FailWith("Expected {context:date} to be after {0}{reason}, but found {1}.", expected,
+                    Subject);
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/>  is not after the specified value.
+        /// </summary>
+        /// <param name="unexpected">The <see cref="DateOnly"/>  that the current value is not expected to be after.</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> NotBeAfter(DateOnly unexpected, string because = "",
+            params object[] becauseArgs)
+        {
+            return BeOnOrBefore(unexpected, because, becauseArgs);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/>  is either on, or after the specified value.
+        /// </summary>
+        /// <param name="expected">The <see cref="DateOnly"/>  that the current value is expected to be on or after.</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> BeOnOrAfter(DateOnly expected, string because = "",
+            params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .ForCondition(Subject.HasValue && Subject.Value.CompareTo(expected) >= 0)
+                .BecauseOf(because, becauseArgs)
+                .FailWith("Expected {context:date} to be on or after {0}{reason}, but found {1}.", expected,
+                    Subject);
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/>  is neither on, nor after the specified value.
+        /// </summary>
+        /// <param name="unexpected">The <see cref="DateOnly"/>  that the current value is expected not to be on nor after.</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> NotBeOnOrAfter(DateOnly unexpected, string because = "",
+            params object[] becauseArgs)
+        {
+            return BeBefore(unexpected, because, becauseArgs);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/> has the <paramref name="expected"/> year.
+        /// </summary>
+        /// <param name="expected">The expected year of the current value.</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> HaveYear(int expected, string because = "", params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .BecauseOf(because, becauseArgs)
+                .WithExpectation("Expected the year part of {context:the date} to be {0}{reason}", expected)
+                .ForCondition(Subject.HasValue)
+                .FailWith(", but found <null>.")
+                .Then
+                .ForCondition(Subject.Value.Year == expected)
+                .FailWith(", but found {0}.", Subject.Value.Year)
+                .Then
+                .ClearExpectation();
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/> does not have the <paramref name="unexpected"/> year.
+        /// </summary>
+        /// <param name="unexpected">The year that should not be in the current value.</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> NotHaveYear(int unexpected, string because = "", params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .BecauseOf(because, becauseArgs)
+                .ForCondition(Subject.HasValue)
+                .FailWith("Did not expect the year part of {context:the date} to be {0}{reason}, but found a <null> DateOnly.", unexpected)
+                .Then
+                .ForCondition(Subject.Value.Year != unexpected)
+                .FailWith("Did not expect the year part of {context:the date} to be {0}{reason}, but it was.", unexpected,
+                    Subject.Value.Year);
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/> has the <paramref name="expected"/> month.
+        /// </summary>
+        /// <param name="expected">The expected month of the current value.</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> HaveMonth(int expected, string because = "", params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .BecauseOf(because, becauseArgs)
+                .WithExpectation("Expected the month part of {context:the date} to be {0}{reason}", expected)
+                .ForCondition(Subject.HasValue)
+                .FailWith(", but found a <null> DateOnly.")
+                .Then
+                .ForCondition(Subject.Value.Month == expected)
+                .FailWith(", but found {0}.", Subject.Value.Month)
+                .Then
+                .ClearExpectation();
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/> does not have the <paramref name="unexpected"/> month.
+        /// </summary>
+        /// <param name="unexpected">The month that should not be in the current value.</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> NotHaveMonth(int unexpected, string because = "", params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .BecauseOf(because, becauseArgs)
+                .WithExpectation("Did not expect the month part of {context:the date} to be {0}{reason}", unexpected)
+                .ForCondition(Subject.HasValue)
+                .FailWith(", but found a <null> DateOnly.")
+                .Then
+                .ForCondition(Subject.Value.Month != unexpected)
+                .FailWith(", but it was.")
+                .Then
+                .ClearExpectation();
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/>  has the <paramref name="expected"/> day.
+        /// </summary>
+        /// <param name="expected">The expected day of the current value.</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> HaveDay(int expected, string because = "", params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .BecauseOf(because, becauseArgs)
+                .WithExpectation("Expected the day part of {context:the date} to be {0}{reason}", expected)
+                .ForCondition(Subject.HasValue)
+                .FailWith(", but found a <null> DateOnly.")
+                .Then
+                .ForCondition(Subject.Value.Day == expected)
+                .FailWith(", but found {0}.", Subject.Value.Day)
+                .Then
+                .ClearExpectation();
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
+        }
+
+        /// <summary>
+        /// Asserts that the current <see cref="DateOnly"/> does not have the <paramref name="unexpected"/> day.
+        /// </summary>
+        /// <param name="unexpected">The day that should not be in the current value.</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> NotHaveDay(int unexpected, string because = "", params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .BecauseOf(because, becauseArgs)
+                .WithExpectation("Did not expect the day part of {context:the date} to be {0}{reason}", unexpected)
+                .ForCondition(Subject.HasValue)
+                .FailWith(", but found a <null> DateOnly.")
+                .Then
+                .ForCondition(Subject.Value.Day != unexpected)
+                .FailWith(", but it was.")
+                .Then
+                .ClearExpectation();
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
+        }
+
+        /// <summary>
+        /// Asserts that the <see cref="DateOnly"/> is one of the specified <paramref name="validValues"/>.
+        /// </summary>
+        /// <param name="validValues">
+        /// The values that are valid.
+        /// </param>
+        public AndConstraint<TAssertions> BeOneOf(params DateOnly?[] validValues)
+        {
+            return BeOneOf(validValues, string.Empty);
+        }
+
+        /// <summary>
+        /// Asserts that the <see cref="DateOnly"/> is one of the specified <paramref name="validValues"/>.
+        /// </summary>
+        /// <param name="validValues">
+        /// The values that are valid.
+        /// </param>
+        public AndConstraint<TAssertions> BeOneOf(params DateOnly[] validValues)
+        {
+            return BeOneOf(validValues.Cast<DateOnly?>());
+        }
+
+        /// <summary>
+        /// Asserts that the <see cref="DateOnly"/> is one of the specified <paramref name="validValues"/>.
+        /// </summary>
+        /// <param name="validValues">
+        /// The values that are valid.
+        /// </param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> BeOneOf(IEnumerable<DateOnly> validValues, string because = "", params object[] becauseArgs)
+        {
+            return BeOneOf(validValues.Cast<DateOnly?>(), because, becauseArgs);
+        }
+
+        /// <summary>
+        /// Asserts that the <see cref="DateOnly"/> is one of the specified <paramref name="validValues"/>.
+        /// </summary>
+        /// <param name="validValues">
+        /// The values that are valid.
+        /// </param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> BeOneOf(IEnumerable<DateOnly?> validValues, string because = "", params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .ForCondition(validValues.Contains(Subject))
+                .BecauseOf(because, becauseArgs)
+                .FailWith("Expected {context:date} to be one of {0}{reason}, but found {1}.", validValues, Subject);
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
+        }
+
+        /// <inheritdoc/>
+        public override bool Equals(object obj) =>
+            throw new NotSupportedException("Calling Equals on Assertion classes is not supported.");
+    }
+}
+
+#endif